requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
pandas==2.2.3
Werkzeug==3.0.6
gunicorn==23.0.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-flask==1.3.0

# Dashboard
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1
numpy==2.2.3

# PDF Generation
reportlab==4.2.5
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.2
